Factor out bicycle.rs
